ACMS-1485: Created new test base class for content admin.
